Adding Barcode reader sample with CameraSource.java
This sample starts with the open source version of CameraSource.java, and
then adds a main activity that demonstrates how to pass data between two
activities.  It also shows how to use the CameraSource to control parameters
such as focus, flash, and zoom.

The first revision of this checkin is the original CameraSource.java, then
the subsequent ones are the one ultimately used in the sample.

Change-Id: I23421e4ea2bd7ce2ea0eedbaece1c37a67f73ba7